requirements.txt
torch
torchvision
tensorboard
pyyaml